review: fix bug in gzip free callbacks
A bug was found upon review, where the gzip free function was passed from the compressor free, as-is, which
meant that the opaque pointer was actually the state pointer. A helper function was added as a callback for
gzip free that properly calls the free function from the original state object.