refactor(utils): GetPropString ensures the json kind is a string before getting the value
The utilities json extension GetPropString() now ensures the propery, if found, is actually a string kind before it attempts to recover the string value. Otherwise the return value remains null. This avoids raising exceptions if the type is not a suitible string which was the original purpose of this function, to safely recover a string value by its name.