refactor!: Correct the spelling of IUnmanagedHeap public interface
BREAKING CHANGE: This change corrects a major typo in the public interface of IUnmanagedHeap. This update does not include any other notable change
and should be safe for upgrading your project safely if previous versions work.

Instructions: You must correct the spelling of any usage of the previous spelling IUnamangedHeap interface or you project will not compile. If dynamic
linking or assembly loading was used, this will cause the module to fail to share types, fail to load.