fix(utils): Fixed a bug that causes inconsistent errors when allocating 0 length blocks. closes #42
This fix introduces new behavior that allows memory blocks of length 0 to make calls down to a native heap without eagerly returning an empty handle. This allows for reallocation if the heap supports it, so a block of length 0 is technically valid, but discouraged for performance reasons. This corrects expected behavior.

This is a difficult UB situation, but vnlib has matured to the point where testing can conclude the allocation of 0 size blocks and allow handles to be created when a heap allows reallocation, otherwise accessing memory that of a block of length 0 is still illegal.